ui: phase 5l — extract Insert mode handler (final)
Move handleInsertMode (~205 lines, the second-largest mode) to
mode_insert.go as a free function. Test shim added.

With this Phase 5l, all 11 per-mode key handlers now live in
per-file mode_*.go and dispatch through the modeHandlers map in
mode_handlers.go.

24/24 ui packages green, vet clean.
